v_20_signal_requirement.txt
pandas
requests
functions-framework
google-cloud-storage
pyarrow
gcsfs